Moving a couple more classes to appropriately named packages
